backend/requirements.txt
djoser
django==2.2.16
django-rest-framework
colorfield
reportlab